feat: migrate model, add user controller and many services

+++ backend/model/booked_slots.go
@@ -8,9 +8,9 @@ import (
 
 type BookedSlot struct {
 	gorm.Model
-	DaycareID int64      `gorm:"not null"`
+	DaycareID uint       `gorm:"not null"`
 	Daycare   PetDaycare `gorm:"foreignKey:DaycareID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
-	PetID     int64      `gorm:"not null"`
+	PetID     uint       `gorm:"not null"`
 	Pet       Pet        `gorm:"foreignKey:PetID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
 	StartDate time.Time
 	EndDate   time.Time
